requirements.txt
torch>=2.1.1,<3.0
transformers>=4.39.0,<5.0.0
asyncio>=3.4.3,<4.0
toml~=0.10.2
pypdf==5.1.0
ratelimit==2.2.1
loguru~=0.7.3
pydantic==2.8.2
tenacity~=8.5.0
rich~=13.9.4
psutil~=6.1.1
python-dotenv~=1.0.1
PyYAML~=6.0.2
docstring_parser==0.16
black>=23.1,<25.0
ruff>=0.0.249,<0.4.5
types-toml>=0.10.8.1
types-pytz>=2023.3,<2025.0
types-chardet>=5.0.4.6
mypy-protobuf>=3.0.0
pytest>=8.1.1
networkx~=3.4.2
aiofiles~=23.2.1
clusterops~=0.1.6
ppft~=1.7.6.9
h11~=0.14.0
numpy~=1.26.4
pillow~=11.1.0
pip~=24.3.1
attrs~=24.3.0
distro~=1.9.0
protobuf~=5.29.2
pyOpenSSL~=24.3.0
cryptography~=44.0.0
filelock~=3.16.1
pox~=0.3.5
ray~=2.40.0
packaging~=23.2
pandas~=2.2.3
grpcio~=1.69.0
requests~=2.32.3
tqdm~=4.67.1
click~=8.1.8
scikit-learn~=1.6.0
scipy~=1.15.0
fsspec~=2024.12.0
aiohttp~=3.11.11
joblib~=1.4.2
opencv-python~=4.10.0.84
numba~=0.60.0
starlette~=0.41.3
onnxruntime~=1.20.1
typer~=0.15.1
fastapi~=0.115.6
uvicorn~=0.34.0
msgpack~=1.1.0
watchfiles~=1.0.3
gradio~=5.10.0
pathspec~=0.12.1
idna~=3.10
multidict~=6.1.0
propcache~=0.2.1
yarl~=1.18.3
aiosignal~=1.3.2
frozenlist~=1.5.0
python-multipart~=0.0.20
aiohappyeyeballs~=2.4.4
uvloop~=0.21.0
jsonschema~=4.23.0
dill~=0.3.9
six~=1.17.0
Pygments~=2.18.0
markdown-it-py~=3.0.0
zipp~=3.21.0
typing_extensions~=4.12.2
sniffio~=1.3.1
platformdirs~=4.3.6
mypy-extensions~=1.0.0
httpcore~=1.0.7
certifi~=2024.12.14
jiter~=0.8.2
mdurl~=0.1.2
setuptools~=75.6.0
pytz~=2024.2
overrides~=7.7.0
charset-normalizer~=3.4.1
cffi~=1.17.1
regex~=2024.11.6
sympy~=1.13.1
mpmath~=1.3.0
decorator~=5.1.1
urllib3~=2.3.0
GPUtil~=1.4.0
MarkupSafe~=2.1.5
ollama~=0.4.5
httpx~=0.27.2
openai~=1.58.1
websockets~=13.1
anyio~=4.7.0
realtime~=2.1.0
pydantic_core~=2.20.1
orjson~=3.10.13
pathos~=0.3.3
multiprocess~=0.70.17
litellm~=1.56.8
sentry-sdk~=2.19.2
pluggy~=1.5.0
Deprecated~=1.2.15
iniconfig~=2.0.0
backoff~=2.2.1
tiktoken~=0.8.0
tokenizers~=0.21.0
importlib_metadata~=8.5.0
Jinja2~=3.1.5
h2~=4.1.0
annotated-types~=0.7.0
langsmith~=0.1.147
requests-toolbelt~=1.0.0
langchain-core~=0.1.53
termcolor~=2.5.0
referencing~=0.35.1
jsonpointer~=3.0.0
jsonschema-specifications~=2024.10.1
SQLAlchemy~=2.0.36
huggingface-hub~=0.27.0
build~=1.2.2.post1
marshmallow~=3.23.2
safetensors~=0.5.0
langchain-community~=0.0.29
cachetools~=5.5.0
accelerate~=1.2.1
soundfile~=0.13.0
av~=12.3.0
faiss-cpu~=1.9.0.post1
librosa~=0.10.2.post1
bitsandbytes~=0.42.0
deprecation~=2.1.0
jsonpatch~=1.33
typing-inspect~=0.9.0
dataclasses-json~=0.6.7
chromadb~=0.6.2
pinecone~=5.4.2
supabase~=2.11.0
pydub~=0.25.1
unique-names-generator~=1.0.2
swarms_memory~=0.1.2
swarm-models~=0.3.0
mysql-connector-python>=8.0.0